[2014/12/25] added      libarray added      libarray/CMakeLists.txt added      libarray/libarray added      libarray/libarray/arrayND.h added      libarray/libarray/CMakeLists.txt added      libarray/libarray/vectorND.h 1. [1ST] Check in the new implementation for ND arrays.
modified   CMakeLists.txt
1. [ADD] Add libarray.

modified   liboptpp/liboptpp/OptValueVector.h
1. [MOD] Correct the type of a loop iterator i from int to size_t.

+++ liboptpp/liboptpp/OptValueVector.h
@@ -71,7 +71,9 @@ namespace OptExt {
 
 		void _SetDests(size_t uNrOfDests, va_list vaDests)
 		{
-			for(int d = 0; d < uNrOfDests; d++) 
+			// MOD-BY-LETEN 2014/12/25:	for(int d = 0; d < uNrOfDests; d++) 
+			for(size_t d = 0; d < uNrOfDests; d++) 
+			// MOD-BY-LETEN 2014/12/25-END
 			{	
 				DT *pDest = va_arg(vaDests, DT*);	
 				VT default_value = va_arg(vaDests, VT);
